backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
groq